+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/entity-type-tabs.tsx
@@ -1,13 +1,12 @@
 import { extractBaseUri, VersionedUri } from "@blockprotocol/type-system-web";
 import { faPlus } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@hashintel/hash-design-system";
-import { Box, Tab, Tabs, Typography } from "@mui/material";
+import { Box, Tab, Tabs, tabsClasses } from "@mui/material";
 import { useRouter } from "next/router";
 import { useEffect, useMemo, useState } from "react";
 import { FRONTEND_URL } from "../../../../lib/config";
 import { EntityTypeDefinitionTab } from "./entity-type-definition-tab";
 import { EntityTypeEntitiesTab } from "./entity-type-entities-tab";
-import { TestTab } from "./tab-button";
 import { TabButton } from "./tab-button";
 import { useEntityType } from "./use-entity-type";
 
@@ -33,7 +32,7 @@ export const EntityTypeTabs = () => {
   }
 
   return (
-    <Box display="flex">
+    <Box display="flex" overflow="visible">
       <Tabs
         value={activeTab}
         onChange={(_, value) => setActiveTab(value)}
@@ -42,7 +41,16 @@ export const EntityTypeTabs = () => {
             height: 3,
             backgroundColor: palette.blue[60],
             minHeight: 0,
+            bottom: -1,
           }),
+        }}
+        sx={{
+          minHeight: 0,
+          overflow: "visible",
+          alignItems: "flex-end",
+          [`.${tabsClasses.scroller}`]: {
+            overflow: "visible !important",
+          },
         }}
       >
         <EntityTypeDefinitionTab value={baseUri} />
@@ -50,23 +58,20 @@ export const EntityTypeTabs = () => {
       </Tabs>
 
       <Box display="flex" ml="auto">
-        {/* <TabButton
-          value="#"
-          sx={(theme) => ({ color: theme.palette.gray[90] })}
-        >
-          <Typography variant="smallTextLabels" sx={{ fontWeight: 500 }}>
-            Create new entity
-          </Typography>
-          <FontAwesomeIcon
-            icon={faPlus}
-            sx={(theme) => ({
-              ...theme.typography.smallTextLabels,
-              color: theme.palette.blue[70],
-              ml: 1,
-            })}
-          />
-        </TabButton> */}
-        <TabButton value="#" label="Create new entity" />
+        <TabButton
+          href="#"
+          label="Create new entity"
+          icon={
+            <FontAwesomeIcon
+              icon={faPlus}
+              sx={(theme) => ({
+                ...theme.typography.smallTextLabels,
+                color: theme.palette.blue[60],
+                ml: 1,
+              })}
+            />
+          }
+        />
       </Box>
     </Box>
   );
